discord_quote.py should only contain bot commands, everything else lives separately
